Add some of the test_cli.py tests
